Background greenlet mode, expanded readme, version bump.
I wanted a mode of running this without having to fork a process and wait a
large amount of time before being able to make requests (I want these things
to run fast), so I introduced a mode where we can run it in a background
green thread much more quickly.

The README also now discusses the compression and character set detection
features, as well as some of the motivation.

Lastly, version bump with these changes.